Taskforce summary tables. Cleanup squadrons.

+++ src/main/java/engima/waratsea/model/ship/Ship.java
@@ -8,6 +8,7 @@ import engima.waratsea.model.ship.data.ShipData;
 import engima.waratsea.model.squadron.Squadron;
 import engima.waratsea.model.taskForce.TaskForce;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -109,7 +110,7 @@ public interface Ship extends PersistentData<ShipData> {
      *
      * @return A map of aircraft types to number of steps of that type.
      */
-    Map<AircraftType, Double> getSquadronSummary();
+    Map<AircraftType, BigDecimal> getSquadronSummary();
 
     /**
      * Get the ship's nationality.
